requirements.txt
djangorestframework>=3.9.2
django>=2.1.7
django_rest_auth>=0.9.3
django_allauth>=0.39.1
psycopg2-binary>=2.7.7
coreapi>=2.3.3
